moving the api functions into api folder
