+++ src/api/buddies.api.js
@@ -1,4 +1,4 @@
-import db from './Config/firebase-setup';
+import db from '../Config/firebase-setup';
 
 const getBuddies = (user, setBuddies) => {
   const buddiesArray = [];
