+++ src/api/tripById.api.js
@@ -13,7 +13,7 @@
 
 // export default getTripById;
 
-import db from './Config/firebase-setup';
+import db from '../Config/firebase-setup';
 
 const getTripById = (currTrip, setCurrTrip) => {
   db.collection('trips')
